+++ lib/dunolint/src/dune0/pps.ml
@@ -31,6 +31,8 @@ module Predicate = struct
   end
 
   module Flag = struct
+    [@@@coverage off]
+
     module Applies_to = struct
       type t =
         [ `any
@@ -49,6 +51,8 @@ module Predicate = struct
   end
 
   module Pp_with_flag = struct
+    [@@@coverage off]
+
     type t =
       { pp : Pp.Name.t
       ; flag : string