+++ lib/test_helpers/src/test_helpers.mli
@@ -40,3 +40,16 @@ val parse
 val is_true : Dunolint.Trilang.t -> unit
 val is_false : Dunolint.Trilang.t -> unit
 val is_undefined : Dunolint.Trilang.t -> unit
+
+(** {1 Linting} *)
+
+(** A helper to run more or less the equivalent of the [lint] command, given a
+    configuration. This runs the dunolint engine in dry-run mode, which would
+    print diffs on the standard output. This is meant to be run from within a
+    call to [Err.For_test.protect], such as, for example:
+
+    {[
+      Err.For_test.protct (fun () -> Test_helpers.run_linter ~config);
+      [%expect {||}]
+    ]} *)
+val run_linter : config:Dunolint.Config.t -> unit