+++ CHANGES.md
@@ -1,3 +1,13 @@
+## 0.0.2025XXXX (unreleased)
+
+### Changed
+
+- Improved testing coverage (#90, @mbarbin).
+
+### Fixed
+
+- Correctly propagate `skip_subtree` when returned from rules (#90, @mbarbin).
+
 ## 0.0.20250804 (2025-08-04)
 
 ### Added